RubikLog/requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django>=4.2.0
django-cors-headers>=4.3.0
djangorestframework>=3.14.0
djongo==1.2.31
dnspython==2.7.0
psycopg2==2.9.10
psycopg2-binary>=2.9.0
pymongo==4.10.1
python-decouple==3.8
pytz==2024.2
sqlparse==0.5.3
numpy>=1.24.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
tensorflow>=2.14.0